feat: create useMediaQuery hook
