+++ package.json
@@ -48,6 +48,7 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.15.8",
+    "@react-spring/types": "^9.3.0",
     "@types/react-helmet": "^6.1.0",
     "@types/styled-components": "^5.1.9",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
